requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
sqlalchemy
psycopg2-binary
python-dotenv
passlib
bcrypt==4.0.1
python-jose[cryptography]
reportlab
requests